support running scripts via symlinks
It's ok that enterShell uses the path as given, because it's just used
to read the contents of the file and for $SCRIPT_FILE.

buildAndRun needs to use the cache for the script, and there's no sense
in making multiple cache entries if there are multiple symlinks to the
same file, so that should use the canonicalized path.

getDerivationTemplateFor definitely needs the canonicalized path in
order to use the file (rather than the symlink itself) in the build.