Avoid misbehaviour due to scripts being built concurrently
It's not safe to assume that `default.nix` in the nix-script cache
directory will only be used by one nix-script instance at a time;
multiple scripts that are started simultaneously could easily all
attempt to write to that file at once. Using a unique temporary
directory for each build avoids this.