pull the overlay out of `eachDefaultSystem`
fixes #76: "error: All overlays passed to nixpkgs must be functions."

I apologize for this oversized PR which actually doesn't change anything
substantial, and only fixes an error. However, when pulling out the
overlay, various problems arouse:

The `pkgs` attribute (and sometimes also the `naersk-lib` attribute) has
to be provided to the functions creating the targets. So those elements
became functions, and I renamed them to `mkNixBla`.

The overlay becomes a bit more complicated because it has to access the
`prev` package set, and call the `mkNixBla` functions with `naersk-lib`.

One positive thing about the overlay being outside is that it can be
used by the `packages` output of the flake itself.

Of course, I am happy to add changes from your side! Maybe you find also
a better solution to this problem!